+++ etc/sql/sql.sql
@@ -43,7 +43,7 @@ CREATE TABLE petitions (
     title STRING INDEX OFF,
     -- relation to cities table
     city STRING,
-    city_answer STRING,
+    city_answer STRING INDEX OFF,
     type STRING,
     description STRING INDEX OFF,
     suggested_solution STRING INDEX OFF,