+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes For IRIS Service
 unreleased
 ==========
 
+2016/08/26 0.0.9
+================
+
  - support event creates supporter data
  - allow admin API to write on petition owner field
  - added city_answer field to petitions